# decision-mesh-api/src/main/resources/META-INF/resources/openapi/internal-apil.yml
openapi: 3.0.3
info:
  title: DecisionMesh Internal Control Plane API
  description: |
    ## Overview
    Internal APIs for DecisionMesh control-plane operations.
    Not intended for public or customer use.
    
    ## Purpose
    These endpoints provide:
    - Decision record retrieval and analysis
    - Policy management and simulation
    - Provider configuration and health monitoring
    - Cost tracking and budget management
    - Analytics and reporting
    - System administration
    
    ## Authentication
    Internal endpoints require elevated privileges and use service-to-service
    authentication with mTLS or internal API keys.
    
    ## Access Control
    - Admin endpoints: Full system access
    - Operator endpoints: Read-only monitoring
    - Analytics endpoints: Data analysis access

  version: 0.2.0-internal
  contact:
    name: DecisionMesh Platform Team
    email: platform-team@decimeshi.com
servers:
  - url: http://localhost:8000/internal
    description: Internal access only (not exposed publicly)
security:
  - InternalApiKey: []
paths:
...
  /decisions/{decision_id}:
    get:
      summary: Retrieve full decision record
      description: Get complete decision record including all metadata, execution details, and outcomes
      operationId: getDecision
      tags:
        - Decisions
      parameters:
        - name: decision_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Decision identifier
      responses:
        "200":
          description: Full decision record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionRecord"
        "404":
          description: Decision not found
...
  /decisions/bulk:
    post:
      summary: Retrieve multiple decision records
      description: Efficient batch retrieval for analytics (max 100 decisions per request)
      operationId: getBulkDecisions
      tags:
        - Decisions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision_ids
              properties:
                decision_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
                  maxItems: 100
      responses:
        "200":
          description: Bulk decision records
          content:
            application/json:
              schema:
                type: object
                properties:
                  decisions:
                    type: array
                    items:
                      $ref: "#/components/schemas/DecisionRecord"
                  not_found:
                    type: array
                    items:
                      type: string
                      format: uuid
  /decisions/search:
    post:
      summary: Advanced decision search
      description: Search decisions with filters, sorting, and pagination
      operationId: searchDecisions
      tags:
        - Decisions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filters:
                  type: object
                  properties:
                    tenant_id:
                      type: string
                    decision_type:
                      type: string
                    status:
                      type: string
                      enum: [success, failure]
                    date_range:
                      type: object
                      properties:
                        start:
                          type: string
                          format: date-time
                        end:
                          type: string
                          format: date-time
                    provider_used:
                      type: string
                    cost_range:
                      type: object
                      properties:
                        min:
                          type: number
                        max:
                          type: number
                sort_by:
                  type: string
                  enum: [created_at, cost_usd, latency_ms]
                  default: created_at
                sort_order:
                  type: string
                  enum: [asc, desc]
                  default: desc
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 20
                offset:
                  type: integer
                  minimum: 0
                  default: 0
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/DecisionRecord"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
  /decisions/purge:
    post:
      summary: Purge decisions older than retention period
      description: Admin-only bulk purge operation for data retention compliance
      operationId: purgeDecisions
      tags:
        - Decisions
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - older_than_days
              properties:
                older_than_days:
                  type: integer
                  minimum: 30
                  description: Purge decisions older than this many days
                tenant_id:
                  type: string
                  description: Optional tenant filter
                dry_run:
                  type: boolean
                  default: true
                  description: If true, return count without deleting
      responses:
        "200":
          description: Purge results
          content:
            application/json:
              schema:
                type: object
                properties:
                  purged_count:
                    type: integer
                  dry_run:
                    type: boolean

  # ==================== POLICY MANAGEMENT ====================
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Internal service API key
  schemas:
    DecisionRecord:
      type: object
      required:
        - decision_id
        - tenant_id
        - status
        - decision_type
        - intent
        - constraints
        - context
        - created_at
      properties:
        decision_id:
          type: string
          format: uuid
        tenant_id:
          type: string
        status:
          type: string
          enum: [pending, executing, success, failure, timeout]
        decision_type:
          type: string
        intent:
          type: object
        constraints:
          type: object
        context:
          type: object
        outcome:
          type: object
          properties:
            result:
              type: object
            metadata:
              type: object
            warnings:
              type: array
        execution_details:
          type: object
          properties:
            total_attempts:
              type: integer
            providers_tried:
              type: array
              items:
                type: string
            fallback_triggered:
              type: boolean
            circuit_breaker_state:
              type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        ttl_expires_at:
          type: string
          format: date-time
...